+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Minor: Moderators can now see which mods start and cancel raids. (#5563)
 - Minor: The emote popup now reloads when Twitch emotes are reloaded. (#5580)
 - Minor: Added `--login <username>` CLI argument to specify which account to start logged in as. (#5626)
+- Minor: Add muting of all application notifications via keybind/window menu. (#5629)
 - Bugfix: Fixed tab move animation occasionally failing to start after closing a tab. (#5426, #5612)
 - Bugfix: If a network request errors with 200 OK, Qt's error code is now reported instead of the HTTP status. (#5378)
 - Bugfix: Fixed restricted users usernames not being clickable. (#5405)